refactor: Restructuring the library to avoid callback problems.
